==> Taobao_v4/01.preprocessing.ipynb <==
import pandas as pd
import numpy as np
import datetime
import pickle

df = pd.read_csv('UserBehavior.csv', names=['uid', 'iid', 'cid', 'type', 'timestamp'])

df['date'] = df['timestamp'].apply(datetime.date.fromtimestamp)

# df['datestr'] = df['date'].apply(str)

df

df.nunique()

np.sort(df['date'].unique())

df_clear = df[(df['date']>=datetime.date(2017, 11, 25)) & (df['date']<=datetime.date(2017, 12, 3))]
df_clear

df_clear['date'].unique()

df_clear.to_csv('UB_clear.csv', index=False)
