==> Taobao_v4/03-1.gensim.ipynb <==
import pandas as pd
import numpy as np
import datetime
import pickle
from tqdm import tqdm

df = pd.read_csv('UB_clear.csv')

df['iid_str'] = df['iid'].apply(str)

df_pv = df[df['type']=='pv']

df_pv

pd.set_option('display.float_format',lambda x : '%.f' % x)
df_pv.describe()

dates = ['2017-11-25'] #, '2017-11-26', '2017-11-27', '2017-11-28', '2017-11-29', '2017-11-30', '2017-12-01' , '2017-12-02', '2017-12-03'

for date in dates:
    df_pv_day = df_pv[df_pv['date']==date]
    df_pv_day_grby_uid = df_pv_day.groupby('uid')
    iid_corpus = []
    not_exist_uid = []
    for _uid in tqdm(range(1018012)):
        try:
            iid_corpus.append(df_pv_day_grby_uid.get_group(_uid).iloc[:, 6].tolist())
        except:
            not_exist_uid.append(_uid)
    with open('./iid_corpus/' + 'iid_corpus_' + date + '.pkl', 'wb') as file:
        pickle.dump(iid_corpus, file)

dates

iid_corpus = []
for date in dates:
    with open('./iid_corpus/' + 'iid_corpus_' + date + '.pkl', 'rb') as file:
        iid_corpus_date = pickle.load(file)
        print(date + ' len: ' + str(len(iid_corpus_date)))
        iid_corpus += iid_corpus_date

len(iid_corpus)

---

import multiprocessing
from gensim.models import Word2Vec

starttime = datetime.datetime.now()

model = Word2Vec(iid_corpus, size=128, window=5, min_count=1, workers=multiprocessing.cpu_count())

endtime = datetime.datetime.now()
print(endtime - starttime)

model.save('iid_1days.model')
