==> Taobao_v4/04.creat_pairs.ipynb <==
import pandas as pd
import numpy as np
import datetime
import pickle
from tqdm import tqdm

dates = ['2017-11-25', '2017-11-26', '2017-11-27', '2017-11-28', '2017-11-29', '2017-11-30',
         '2017-12-01', '2017-12-02', '2017-12-03']

from itertools import tee, islice, zip_longest
def get_next(some_iterable, window=1):
    items, nexts = tee(some_iterable, 2)
    nexts = islice(nexts, window, None)
    return zip_longest(items, nexts)

for date in dates:
    iid_pairs = []
    with open('./iid_corpus/' + 'iid_corpus_' + date + '.pkl', 'rb') as file:
        iid_corpus_date = pickle.load(file)
    for cs in tqdm(iid_corpus_date):
        for target, _next in get_next(cs):
            if _next is not None:
                iid_pairs.append((target, _next))
    with open('./iid_pairs/' + 'iid_pairs_' + date + '.pkl', 'wb') as file:
        pickle.dump(iid_pairs, file)
