==> Taobao_v4/05.creat_table.ipynb <==
import pandas as pd
import numpy as np
import datetime
import pickle
from tqdm import tqdm
from gensim.models import Word2Vec

df = pd.read_csv('UB_clear.csv')

df_pv = df[df['type']=='pv']

pd.set_option('display.float_format',lambda x : '%.f' % x)
df_pv.describe()

iid_uni = df_pv['iid'].unique()

len(iid_uni)

iid_uni.sort()

iid_uni[:10]

model = Word2Vec.load("iid_1days.model")

model.wv.most_similar(positive='1', topn=99)

PART = '1'

dict_knn_table = {}
iid_not_in_vocab = []

for iid in tqdm(iid_uni[0:1000000]):
    try:
        dict_knn_table[str(iid)] = model.wv.most_similar(positive=str(iid), topn=99)
    except:
        iid_not_in_vocab.append(iid)

with open('./dict_knn_table/' + 'knn_table_' + PART + '.pkl', 'wb') as file:
    pickle.dump(dict_knn_table, file)

with open('./dict_knn_table/' + 'iid_not_in_vocab_' + PART + '.pkl', 'wb') as file:
    pickle.dump(iid_not_in_vocab, file)

# PART = '3'
# ...
# PART = '4'
# ...
