==> Taobao_v4/06-1.knn_rank_9days.ipynb <==
import pandas as pd
import numpy as np
import datetime
import pickle
from tqdm import tqdm
from gensim.models import Word2Vec

model = Word2Vec.load("./iid_1days.model")

knn_tables = []
for part in range(1, 5):
    with open('./dict_knn_table/knn_table_'+ str(part) +'.pkl', 'rb') as file:
        knn_tables.append(pickle.load(file))

dates = ['2017-11-25', '2017-11-26', '2017-11-27', '2017-11-28', '2017-11-29', '2017-11-30',
         '2017-12-01', '2017-12-02', '2017-12-03'] # 

df_knn_rank = []

for idx, date in tqdm(enumerate(dates)):
    with open('./iid_pairs/' + 'iid_pairs_' + date + '.pkl', 'rb') as file:
        iid_pair_date = pickle.load(file)

    df_knn_rank.append(pd.DataFrame(iid_pair_date, columns=['target', 'next']))

    def ranking(r):
        for part in range(4):
            if r.target in knn_tables[part]:
                lt = knn_tables[part][r.target]
                ids, cos = zip(*lt)
                try:
                    rank = ids.index(r.next) + 1
                except:
                    rank = np.nan
                break
        else:
            rank = np.nan
        return rank

    df_knn_rank[idx]['rank'] = df_knn_rank[idx].apply(ranking, axis=1)

with open('./knn_result/lt_df_knn_rank_9days.pkl', 'wb') as file:
    pickle.dump(df_knn_rank, file)

---

pd.set_option('display.float_format',lambda x : '%.2f' % x)

df_knn_rank[0].describe()

df_knn_rank[1].describe()

df_knn_rank[2].describe()

df_knn_rank[3].describe()

df_knn_rank[4].describe()

df_knn_rank[5].describe()

df_knn_rank[6].describe()

df_knn_rank[7].describe()

df_knn_rank[8].describe()
