==> Taobao_v4/08-1.resorting_table.ipynb <==
import pandas as pd
import numpy as np
import datetime
import time
import pickle
from tqdm import tqdm
import matplotlib.pyplot as plt
%matplotlib inline
from collections import Counter
from random import choices, sample
from gensim.models import Word2Vec
import gc
import keras

PART = '1'

with open('./dict_knn_table/' + 'knn_table_' + PART + '.pkl', 'rb') as file:
    dict_table = pickle.load(file)

model_w2v = Word2Vec.load("iid_1days.model")

model = keras.models.load_model('./saved_model/2017-11-25_8666470_0423.h5')

for iid in tqdm(list(dict_table.keys())):
    topn, cos = zip(*dict_table[iid])
    score = [ float(model.predict(np.hstack(( model_w2v.wv[iid], model_w2v.wv[n] )).reshape(1, 256))[0][0]) for n in topn ]
    new_rank = list(zip(topn, score))
    new_rank.sort(key=lambda x : x[1], reverse=True)
    dict_table[iid] = new_rank

with open('./dict_NN_table/' + 'NN_table_' + PART + '.pkl', 'wb') as file:
    pickle.dump(dict_table, file)

PART = '2'

with open('./dict_knn_table/' + 'knn_table_' + PART + '.pkl', 'rb') as file:
    dict_table = pickle.load(file)

for iid in tqdm(list(dict_table.keys())):
    topn, cos = zip(*dict_table[iid])
    score = [ float(model.predict(np.hstack(( model_w2v.wv[iid], model_w2v.wv[n] )).reshape(1, 256))[0][0]) for n in topn ]
    new_rank = list(zip(topn, score))
    new_rank.sort(key=lambda x : x[1], reverse=True)
    dict_table[iid] = new_rank

with open('./dict_NN_table/' + 'NN_table_' + PART + '.pkl', 'wb') as file:
    pickle.dump(dict_table, file)

# PART = '3'
# ...
# PART = '4'
# ...
