==> Taobao_v4/09.NN_rank.ipynb <==
import pandas as pd
import numpy as np
import datetime
import time
import pickle
from tqdm import tqdm
import matplotlib.pyplot as plt
%matplotlib inline
from collections import Counter
from random import choices, sample
from gensim.models import Word2Vec
import gc
pd.set_option('display.float_format',lambda x : '%.2f' % x)

NN_tables = []
for part in range(1, 5):
    with open('./dict_NN_table/NN_table_'+ str(part) +'.pkl', 'rb') as file:
        NN_tables.append(pickle.load(file))

with open('./knn_result/lt_df_knn_rank_9days.pkl', 'rb') as file:
    lt_df_knn_rank = pickle.load(file)

for i in range(len(lt_df_knn_rank)):
    lt_df_knn_rank[i].dropna(inplace=True)

lt_df_knn_rank[0]

def ranking(r):
    for part in range(4):
        if r.target in NN_tables[part]:
            lt = NN_tables[part][r.target]
            ids, score = zip(*lt)
            try:
                rank = ids.index(r.next) + 1
            except:
                rank = np.nan
            break
    else:
        rank = np.nan
    return rank

for i in tqdm(range(len(lt_df_knn_rank))):
    lt_df_knn_rank[i]['NN_rank'] = lt_df_knn_rank[i].apply(ranking, axis=1)

with open('./NN_result/lt_df_NN_rank_9days.pkl', 'wb') as file:
    pickle.dump(lt_df_knn_rank, file)
